unified-site/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
cohere==5.11.0
qdrant-client==1.12.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
python-multipart==0.0.20
sqlalchemy==2.0.36
asyncpg==0.30.0
langchain-community==0.3.0
langchain-core==0.3.0
openai==1.52.0
tiktoken==0.8.0